chore(tags): move 1.x deprecated functionality back to 1.x
