+++ pkg/sendconfig/sendconfig_test.go
@@ -128,51 +128,3 @@ func Test_updateReportingUtilities(t *testing.T) {
 	assert.True(t, hasSHAUpdateAlreadyBeenReported([]byte("yet-another-fake-sha")))
 	assert.True(t, hasSHAUpdateAlreadyBeenReported([]byte("yet-another-fake-sha")))
 }
-
-func Test_getIngressControllerTags(t *testing.T) {
-	type args struct {
-		config Kong
-	}
-	tests := []struct {
-		name string
-		args args
-		want []string
-	}{
-		{
-			name: "configuration with tag support and filter tags",
-			args: args{
-				config: Kong{
-					DeprecatedHasTagSupport: true,
-					FilterTags:              []string{"foo-tag", "bar-tag"},
-				},
-			},
-			want: []string{"foo-tag", "bar-tag"},
-		},
-		{
-			name: "configuratiion with tag support and no filter tags",
-			args: args{
-				config: Kong{
-					DeprecatedHasTagSupport: true,
-					FilterTags:              []string{},
-				},
-			},
-			want: nil,
-		}, {
-			name: "configuration with no tag support",
-			args: args{
-				config: Kong{
-					DeprecatedHasTagSupport: false,
-				},
-			},
-			want: nil,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got := GetIngressControllerTags(tt.args.config)
-			if !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("getIngressControllerTags() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
